Re-organized the plumed module
