+++ deep_cartograph/modules/plumed/__init__.py
@@ -1,1 +1,5 @@
+from .cli import *
+from .command import *
+from .input_file import *
+from .colvars import *
 from .utils import *
