--- deep_cartograph/modules/plumed/command/__init__.py
@@ -1,1 +1,0 @@
-from.command import *
