--- deep_cartograph/modules/plumed/input_file/__init__.py
@@ -1,1 +1,0 @@
-from .input_file import *
