+++ deep_cartograph/modules/plumed/utils.py
@@ -1,11 +1,7 @@
 # Import modules
-import io 
+
 import os
-import sys
 import logging
-import subprocess
-import numpy as np
-import pandas as pd
 from pathlib import Path
 
 # Import local modules
@@ -16,242 +12,6 @@ logger = logging.getLogger(__name__)
 
 # Set constants
 DEFAULT_FMT = '%14.10f'
-
-# I/O CSV Handlers
-# ------------
-#
-# Functions to handle PLUMED CSV files
-def read_as_pandas(colvars_path: str) -> pd.DataFrame:
-    '''
-    Function that reads a COLVARS file and returns a pandas DataFrame with the same column names as in the COLVARS file.    
-    The time column in ps will be converted to ns.
-
-    If the logger level is set to DEBUG, information about the column names will be printed.
-
-    Inputs
-    ------
-
-        colvars_path    (str):          COLVARS file path
-
-    Outputs
-    -------
-
-        colvars_df      (pandas DataFrame):      COLVARS data
-    '''
-
-    # Read column names
-    column_names = read_column_names(colvars_path)
-
-    # Read COLVARS file
-    colvars_df = pd.read_csv(colvars_path, sep='\s+', dtype=np.float32, comment='#', header=None, names=column_names)
-
-    # Convert time from ps to ns - working with integers to avoid rounding errors
-    colvars_df["time"] = colvars_df["time"] * 1000 / 1000000
-
-    # Show info of traj_df
-    writtable_info = io.StringIO()
-    colvars_df.info(buf=writtable_info)
-    logger.debug(f"{writtable_info.getvalue()}")
-
-    return colvars_df
-
-def read_column_names(colvars_path: str) -> list:
-    '''
-    Reads the column names from a COLVARS file. 
-
-    Inputs
-    ------
-
-        colvars_path    (str):          COLVARS file path
-
-    Outputs
-    -------
-
-        column_names    (list of str):  list with the column names
-    '''
-
-    # Read first line of COLVARS file
-    with open(colvars_path, 'r') as colvars_file:
-        first_line = colvars_file.readline()
-
-    # Separate first line by spaces
-    first_line = first_line.split()
-
-    # The first element is "#!" and the second is "FIELDS" - remove them
-    column_names = first_line[2:]
-
-    return column_names
-
-def write_as_csv(dataframe, path):
-    """
-    Writes a pandas DataFrame to a CSV file keeping the same column names but in PLUMED format.
-    Note that the time column is assumed to be in ns and will be converted to ps!
-
-    If the file already exists, it will append the new data to the existing file.
-
-    Inputs
-    ------
-
-        dataframe   (pandas DataFrame):    DataFrame to be written
-        path                     (str):    path to the CSV file including the file name
-
-    """
-
-    # Convert time from ns to ps
-    dataframe["time"] = dataframe["time"] * 1000
-
-    # Check if file already exists
-    if not os.path.isfile(path):
-
-        # Create csv file with header line
-        header_line = "#! FIELDS " + " ".join(dataframe.columns)
-        with open(path, 'w') as csv_file:
-            csv_file.write(header_line + "\n")
-    
-    else:
-        # Erase first dataframe row and add time offset
-        # Find last time in csv file
-        with open(path, 'r') as csv_file:
-            last_line = csv_file.readlines()[-1]
-        last_time = float(last_line.split()[0])
-
-        # Erase first row - repeated sample for same initial conditions 
-        dataframe = dataframe.drop(dataframe.index[0])
-
-        # Add time offset to dataframe
-        dataframe["time"] = dataframe["time"] + last_time
-
-    # Close csv file
-    csv_file.close()
-
-    # Append data to csv file
-    dataframe.to_csv(path, mode='a', header=False, index=False, sep=' ', float_format='%.6f')
-
-    return
-
-
-# PLUMED driver
-# -------------
-#
-# Returns the corresponding PLUMED driver shell command
-def get_driver_command(plumed_input: str, traj_path: str = None, topology_path: str = None):
-    '''
-    Function that creates a PLUMED DRIVER Shell command. It returns the command as a string
-
-    Example:
-
-        "driver --plumed plumed_input --ixyz traj_path --natoms num_atoms"
-
-    Inputs
-    ------
-
-        plumed_input     (str):              PLUMED input file path
-        traj_path        (str):              path to trajectory file
-        topology_path    (str):              path to topology file
-
-    Outputs
-    -------
-
-        driver_command   (str):              PLUMED DRIVER command
-    '''
-
-    # Initialize
-    driver_command = []
-        
-    # Add driver flag
-    driver_command.append("driver")
-
-    # Add plumed flag
-    driver_command.append("--plumed")
-
-    # Make sure plumed input is given with the absolute path
-    plumed_input = os.path.abspath(plumed_input)
-
-    # Add plumed input
-    driver_command.append(plumed_input)
-
-    # Add trajectory or noatoms flag
-    if traj_path is None:
-        driver_command.append("--noatoms")
-    else:
-        traj_flag = get_traj_flag(traj_path)
-        driver_command.append(traj_flag)
-        
-        if Path(traj_path).suffix == ".pdb":
-            traj_path = check_CRYST1_record(traj_path, Path(topology_path).parent)
-            
-        traj_path = os.path.abspath(traj_path)
-        driver_command.append(traj_path)
-
-    # Find the number of atoms if topology is given (some traj formats do not need this)
-    if topology_path is not None:
-        num_atoms = md.get_number_atoms(topology_path)
-        driver_command.append("--natoms")
-        driver_command.append(str(num_atoms))
-
-    # Join command
-    driver_command = " ".join(driver_command)
-
-    return driver_command 
-
-def run_plumed(plumed_command: str, plumed_settings: dict, plumed_timeout: int) -> None:
-    """
-    Runs PLUMED through command line, setting up the necessary environment variables and modules.
-
-    Inputs
-    ------
-
-        plumed_command  (str):               PLUMED command to execute (See Command Line Tools in PLUMED manual)
-        plumed_settings (dict):              (Optional) Settings for PLUMED (binaries, kernel, etc.)
-        plumed_timeout  (int):               (Optional) timeout for PLUMED in seconds
-    
-    Returns
-    -------
-    
-        tuple: (stdout, stderr) from the PLUMED execution
-    """
-
-    all_commands = []
-    plumed_binary = plumed_settings.get('bin_path', 'plumed') if plumed_settings else 'plumed'
-    
-    if plumed_settings:
-        if plumed_settings.get('env_commands'):
-            all_commands.append(" && ".join(plumed_settings.get('env_commands')))
-        
-        if plumed_settings.get('kernel_path'):
-            os.environ['PLUMED_KERNEL'] = plumed_settings.get('kernel_path')
-    
-    all_commands.append(f"{plumed_binary} {plumed_command}")
-    command_str = " && ".join(all_commands)
-    
-    logger.info(f"Executing PLUMED command: {command_str}")
-
-    try:
-        # Execute PLUMED redirecting output to the log file
-        completed_process = subprocess.run(args=command_str, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, timeout=plumed_timeout, text=True) 
-        
-        stdout, stderr = completed_process.stdout, completed_process.stderr
-        
-        if logger.isEnabledFor(logging.DEBUG):
-            # Send standard output to the log file 
-            logger.info(stdout)
-            
-        # Check if PLUMED failed
-        if completed_process.returncode != 0:
-            logger.error("PLUMED execution failed! \n")
-            logger.error(stderr)
-            sys.exit(1)
-            
-        return stdout, stderr
-    
-    except subprocess.TimeoutExpired:
-        logger.error("PLUMED execution timed out!")
-        return None, "TimeoutExpired"
-    
-    except Exception as e:
-        logger.error(f"Unexpected error: {e}")
-        return None, str(e)
-
 
 # Get labels for features/cvs
 # ---------------------------
